% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_function.R
\name{new_function}
\alias{new_function}
\title{Matches the current get_play_data functionality for setting binary indicators.
all functions and variables that are created have the prefix new_}
\usage{
new_function(new_play)
}
\arguments{
\item{new_play}{list containing data for the play. This is the same as the play parameter currently taken by get_play_data}
}
\description{
Matches the current get_play_data functionality for setting binary indicators.
all functions and variables that are created have the prefix new_
}
