% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_and_tests.R
\name{test_new_function_validity}
\alias{test_new_function_validity}
\title{runs the new code and the current code for every available play
the current code is copy and pasted into a function that
takes play as an input and returns a list of the data it creates
the outside data it needs(nfl_stat_map) is provided for it}
\usage{
test_new_function_validity(number_of_games = 25, smoke_test = FALSE)
}
\arguments{
\item{number_of_games}{number of random games to test. not used if smoke_test is true}

\item{smoke_test}{TRUE/FALSE when true, run every available game}
}
\description{
runs the new code and the current code for every available play
the current code is copy and pasted into a function that
takes play as an input and returns a list of the data it creates
the outside data it needs(nfl_stat_map) is provided for it
}
