% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_and_tests.R
\name{timing_summary}
\alias{timing_summary}
\title{utility for viewing and summarizing timing data that is recorded
#TODO clear ticlog and save only the number of timings and total time for each event}
\usage{
timing_summary()
}
\description{
utility for viewing and summarizing timing data that is recorded
#TODO clear ticlog and save only the number of timings and total time for each event
}
